Remove python3 from Spack environment for Myriad
I added Python 3 for hpgmg, which however works also with Python 2.  The problem
of Python 3 on Myrias is that it doesn't set the rpath, and at some point of the
build process in Spack the `python3.8` executable can't find its own
`libpython`, because at that point the loader would be using only default system
search path (hint: use `LD_DEBUG=all` to see this).  Using Python 2 solves the
issue because it's installed in `/usr`.