Lie about version number of libtool spec in Cosma8 Spack environment
`libtool@2.4.2` is deprecated in Spack, but it's the system version and system
libraries have been built with that version of libtool, and we need to match it
when building libraries/programs linking to sysem libraries.  The trick uses
here is to lie about what version of `libtool` is available in the system, so
that Spack won't skip it.  libtool is a bit of a mess.