Remove unused programming environments
We don't use programming environments in ReFrame, the choice of
compiler is made by spack and controlled by te spack spec. The
available compilers on each system are recorded in their spack
environments. Setting them in ReFrame does nothing, because
spack will not pay attention to the ReFrame programming environment