[cosma8] Fix oneAPI compiler settings in Spack environment
The main problem was that the compiler itself links to `libimf.so` (Intel Math
Library) and relies on `LD_LIBRARY_PATH` being set correctly for the dynamic
loader to find it at runtime.  Likewise, the compiler generates binaries which
link to `libimf.so` without setting rpath/runpath, and so `LD_LIBRARY_PATH`
would be needed again to use those binaries.

The solution was to update the compiler settings in the spack environment to add
the directory where `libimf.so` is to `LD_LIBRARY_PATH`, to make the compiler
work in the first place, _and_ to list the same directory as an additional rpath
to set when linking objects with oneAPI toolchains.