CSD3: Modify ReFrame configuration
Always request all of the memory on a node as `--exclusive` does not
imply this according to the current SLURM documentation
(https://slurm.schedmd.com/sbatch.html#OPT_exclusive). This can make a
difference to benchmark results when not using all of the cores on a
node due to the increased memory bandwidth available.

Also increase the job_submit_timeout as the SLURM controller can be a
bit slow on CSD3.