Update CSD3 Spack environments
Following the upgrade of the Cascade Lake nodes to Rocky Linux 8, all
CPU partitions can use the same software stack. We switch to using the
software stack built for the Cascade Lake nodes as this is the newest
and works on all partitions.

Instead of adding system provided packages as externals, we instead
chain our Spack environment to that of the system Spack instance.